fix: SendMessage/PostMessage declared WPARAM/LPARAM as 32-bit DWORD (#203)
WPARAM/LPARAM are pointer-sized (UINT_PTR/LONG_PTR). The DWORD (uint32)
declaration zero-extended negative values on x64 - the built-in
monitoron (lParam=-1) became 0x00000000FFFFFFFF. Redeclared both as
nint (SendMessage now returns IntPtr, matching LRESULT; PostMessage
keeps its BOOL return) and sign-extended the stored ints at the call
sites in SendMessageCommand and MainWindow. Scoped narrowly per #203;
the broader interop consolidation is #210.

+++ src/Commands/SendMessageCommand.cs
@@ -86,7 +86,8 @@ public class SendMessageCommand : Command {
                     }
                     else {
                         Logger.Instance.Log4.Info($"{this.GetType().Name}: SendMessage(\"{win.MainWindowTitle}\", {Msg}, {WParam}, {LParam}) - {ToString()}");
-                        Win32.SendMessage(win.MainWindowHandle, (DWORD)Msg, (DWORD)WParam, (DWORD)LParam);
+                        // #203: (nint) sign-extends the stored int (lParam=-1 stays -1 on x64).
+                        Win32.SendMessage(win.MainWindowHandle, (DWORD)Msg, (nint)WParam, (nint)LParam);
                     }
                 }
                 else {
@@ -97,7 +98,7 @@ public class SendMessageCommand : Command {
             else {
                 IntPtr h = Win32.GetForegroundWindow();
                 Logger.Instance.Log4.Info($"{this.GetType().Name}: SendMessage(<forground window>, {Msg}, {WParam}, {LParam}) - {ToString()}");
-                Win32.SendMessage(h, (DWORD)Msg, (DWORD)WParam, (DWORD)LParam);
+                Win32.SendMessage(h, (DWORD)Msg, (nint)WParam, (nint)LParam);
             }
         }
         catch (Exception e) {
